Test: Fix encrypted log test by replacing count down latch with turbine
FYI.1: It is quite hard to work with this test because 'TooManyRequests'
can be hit quite easily, at which point you need to wait for a few
minutes (15' minutes to half an hour) before being able to verify that
the test is working as expected.

FYI.2: Also, in order to verify that the test is working as expected one
would need to use the debugger and navigate through the flow manually.
Otherwise the test might success with a false positive, but actually
silently failing on (at least) this 'TooManyRequests' problem mentioned
above.