Migration: Replace date created with synthetic primary key
FYI: This is done in order to get rid 'DateTimeUtils' and everything
related to handling 'dateCreated' just for ordering purposes. Using a
synthetic auto-generated primary key is enough and will achieve the same
ordering of encrypted logs, but this time, without depending on the date
that encrypted log was created.

PR Discussions:
- https://github.com/Automattic/EncryptedLogging/pull/17
#discussion_r2120752334
- https://github.com/Automattic/EncryptedLogging/pull/18
#discussion_r2149442437

Related Issue: Consider removing EncryptedLog#dateCreated and
EncryptedLogModel#dateCreated fields #12
- https://github.com/Automattic/EncryptedLogging/issues/12