Tmp: Don't upload encrypted log, just handle the encryption part
FYI: This change is for debugging purposes so that to avoid uploading
log files that are too big, and until we figure out how big a log file
should be to cause this OOM problem. Then, to try to fix the problem and
see if 'logEncrypter.encrypt(...)' could now handle such big log files.