Fix: Add logs size limit and pre-trip logs before encrypting
FYI: This 5MB max logs size limit is both arbitrary and empirical
because it seems that for (at least) some standard devices, large logs
files of 100MB+ can easily cause OOM. Also, this 5MB max logs size limit
is chosen because the API accepts logs up to 10MB, but we leave enough
headroom for encoding overhead.

PS: It is worth to note that we could apply a different kind of a
solution (better one), like streaming encryption or chunked processing.
But, I don't think it is worth it, nor I think it is worth sending such
large log files to the API.